requirements.txt
albumentations
einops
numpy
opencv-python
PyYAML
torch
torchinfo
tqdm
wandb
